Moved IO wrappers and heredoc eof warning displays proper linono
